添加了材料搜索框

+++ src/cn/gavin/save/SaveHelper.java
@@ -85,7 +85,7 @@ public class SaveHelper {
         editor.putLong("hitRate", heroN.getHitRate());
         editor.putFloat("parry", heroN.getParry());
         editor.putFloat("dodgeRate", heroN.getDodgeRate());
-        editor.putFloat("clickPointAward", heroN.getClickPointAward());
+        editor.putLong("clickPointAward", heroN.getClickPointAward());
         editor.apply();
     }
